requirements.txt
# Core dependencies
flask>=3.0.0
flask_sqlalchemy==3.0.4
SQLAlchemy==1.4.52
flask-cors>=4.0.0
pymysql>=1.1.0
requests>=2.31.0
python-dotenv>=1.0.1
pillow>=10.2.0
openpyxl>=3.1.2

# Development dependencies
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.3.0
pre-commit>=3.3.0
pytest>=7.4.0
pytest-cov>=4.1.0
